fix: Encoding of List offsets was incorrect when slice offsets begin with zero
When encoding offsets the code had an optimization to reuse the offsets if the first offset was zero assuming the slice already pointed to first element.  But the offset can also be zero if all previous lists were empty.  When this occured it mold make all lists in the slice as empty, even if they shouldn't be.